Support description item alteration addendums in expandable self effect actions
