+++ src/styles/legacy/_chat.scss
@@ -81,19 +81,6 @@
         }
     }
 
-    .addendum {
-        @include frame-elegant;
-        background: rgba($alt-dark, 0.1);
-        border-radius: 2px;
-        color: var(--color-pf-alternate-darker);
-        padding: var(--space-4);
-
-        h4 {
-            margin: 0 0 var(--space-4) 0;
-            text-align: center;
-        }
-    }
-
     .card-buttons {
         display: flex;
         flex-basis: 100%;
