+++ src/styles/ui/sidebar/chat/_index.scss
@@ -100,6 +100,20 @@
     > .message-content {
         @import "conditions";
 
+        .addendum {
+            @include frame-elegant;
+            background: rgba($alt-dark, 0.1);
+            border-radius: 2px;
+            color: var(--color-pf-alternate-darker);
+            padding: var(--space-4);
+
+            h4 {
+                font-weight: 600;
+                margin: 0 0 var(--space-4) 0;
+                text-align: center;
+            }
+        }
+
         .dice-roll {
             .dice-result > [data-visibility="gm"] {
                 background: var(--visibility-gm-bg);
